cleaned up main.py and refactored code.
