+++ skunkbooth/main.py
@@ -1,12 +1,9 @@
 import logging
-from datetime import datetime
 from functools import partial
 from multiprocessing import Process, Queue
 from time import time
-from typing import List, Tuple
 
 from asciimatics.effects import Print
-from asciimatics.event import Event, KeyboardEvent
 from asciimatics.exceptions import ResizeScreenError, StopApplication
 from asciimatics.scene import Scene
 from asciimatics.screen import Screen
@@ -19,6 +16,9 @@ from skunkbooth.utils.frames import (
 )
 from skunkbooth.utils.videoManager import videoManager
 from skunkbooth.utils.webcam import Webcam
+from skunkbooth.utils.helpers import (
+    global_shortcuts, toggleFlag, CamDimensions 
+)
 
 # Initialize logger
 logging.basicConfig(
@@ -28,35 +28,12 @@ logging.basicConfig(
     format='%(asctime)s,%(msecs)d %(levelname)-8s [%(filename)s:%(lineno)d] %(message)s'
 )
 
-
-def global_shortcuts(event: Event) -> None:
-    """Event handler for global shortcuts"""
-    ctrlQCode = Screen.ctrl('q')
-    ctrlWCode = Screen.ctrl('w')
-    if isinstance(event, KeyboardEvent):
-        c = event.key_code
-        # Stop on q, esc, ctrl+q and ctrl+w
-        if c in (Screen.KEY_ESCAPE, ord('q'), ctrlQCode, ctrlWCode):
-            raise StopApplication("User pressed quit")
-
-
 def main() -> None:
     """Main driver function"""
     # Video saving
     vidBuf = Queue(32767)
     vid = Process(target=videoManager, args=[vidBuf])
     vid.start()
-
-    def toggleFlag(flag: List[int]) -> None:
-        """Temp function for toggling video recording from inside screen"""
-        flag[0] = not flag[0]
-        # re-initialize VideoIO for new file name
-        if flag[0]:
-            VID_FILE = f"{PIC_DIR}/Video-{datetime.now().strftime('%Y-%m-%d-%H-%M-%S')}.avi"
-            logging.info(f"Recording new video - {VID_FILE}")
-            vidBuf.put(VID_FILE)
-        else:
-            logging.info("Recording stopped.")
 
     TOP_MARGIN = 4
     image_selection = ImageSelectionModel()
@@ -67,23 +44,10 @@ def main() -> None:
     logging.info(
         "Screen initialized Height:{} Width:{}".format(screen.height-8, screen.width)
     )
-
+    
     # last_scene = None
     filters = filterManager()
     converter = Blocks(screen.height, screen.width, uni=True, fill_background=True)
-
-    def CamDimensions(height: int, width: int) -> Tuple[int, int, int]:
-        """Calculate dimensions for vertical squeeze screen sizes"""
-        if width / height >= 4:
-            height -= 8
-            var_dim = int(height * 4)  # Max width is around twice height in most cases
-            offset = int(width / 2 - var_dim / 2.5 - width / 5)
-            return (height, var_dim, offset)
-        # Add margins of 1/6x,y if no vertical squeeze
-        height = int(height * 2 / 3)
-        width = int(width * 2 / 3)
-        return (height, width, 2)
-
     (webcam_height, webcam_width, offset) = CamDimensions(screen.height, screen.width)
 
     logging.info(
